Fix for logout event race condition, not an idea fix
Fix is to add an argument to SessionService#checkAuthentication(loggingOut)
The flag indicates to checkAuthentication to not trigger auth confirmed when checking authentication.
This in turn completely bypasses issues with having two conflicting events fired (in the correct order even) but have clients receive them in the incorrect order.
Which meant that clients would see a logout event followed by a login event, which would mean you would see the logout screen for a fraction of a second then back to home screen