UI-Router resolve issues fixed
Using ui-router resolve to enforce session status for certain states
is good, but my original implementation was naive and based upon
invalid angular.js knowledge.
What I incorrectly did was create a service that would query auth
status, which in and of its self... worked.
The misunderstanding was due to the fact that in angular,
services are singletons.
This meant that the service would always return the *same* value
as the first time it ran. Meaning it was not possible to login if
the login service had run and returned false.
The system would be logged in behind the scenes and have a proper
session but resolve would block access to logged in states.
This was fixed by using functions instead of services for ui-router
to resolve. The functions would obtain the backing services at
run time (due to the fact that services cannot be injected during
config) and then queried. This means that the result would *not*
be cached and will always work properly